+++ src/Log/Engine/ConsoleEngine.php
@@ -48,12 +48,12 @@ class ConsoleEngine extends BaseEngine
       * @param string $level e.g debug, info, notice, warning, error, critical, alert, emergency.
       * @param string $message 'this is a {what}'
       * @param array $context  ['what'='string']
-      * @return bool
+      * @return void
       */
-    public function log(string $level, string $message, array $context = []) : bool
+    public function log(string $level, string $message, array $context = []) : void
     {
         $message = $this->format($level, $message, $context);
 
-        return (bool) $this->output->write("<{$level}>{$message }</{$level}>");
+        $this->output->write("<{$level}>{$message }</{$level}>");
     }
 }