+++ src/Log/Engine/EmailEngine.php
@@ -85,14 +85,14 @@ class EmailEngine extends BaseEngine
       * @param string $level e.g debug, info, notice, warning, error, critical, alert, emergency.
       * @param string $message 'this is a {what}'
       * @param array $context  ['what'='string']
-      * @return bool
+      * @return void
       */
-    public function log(string $level, string $message, array $context = []) : bool
+    public function log(string $level, string $message, array $context = []) : void
     {
         $message = $this->format($level, $message, $context) . "\n";
         $subject = 'Log: ' . strtoupper($level);
 
-        return $this->send($subject, $message);
+        $this->send($subject, $message);
     }
 
     /**